Update README.md
Due to the way that we handled our driving and navigating, and especially how we managed our repo my commit history is significantly skewed on the project. I may have a large number of commits but the reason is because everyone pushed changes to my branches and was in a live server with me for the last couple of days of development when we did the most work. Despite my large contributions overall, GitHub simply isn't able to track how Darran and Kawika really did so much more impressive work than a few commits and they brought the project together. After picking apart huge commits full of a huge day's work more than half of our project and vision goes to Darran and Kawika who made our project into an experience, and our ideas and models into a game. They were an excellent team, they were trustworthy, passionate, and consistent. My greatest contributions were in features that were stretch goals, and pair programming with my teammates. They are the ones who brought the vision to life and I hope they can see it the same way. That being said, I specifically integrated Socket.io, Auth0, and laid a foundation for my team to build from in the form of creating some game-logic as they were unfamiliar with the process, and creating some request flows like "this is how we should get our authorization from Auth0, send it to the database, and then update our client with new info" They figured out the rest and helped me every step of the way while they did it! They were EXCELLENT teammates. We were constantly in agreement not because our ideas were perfect, but because they had the will and maturity to make it work whether it be by steaming ahead or throwing everything overboard to ensure our success before we flooded our boat. 